# hottub/src/main/resources/openapi/hottub-api-v1.0.yml
openapi: 3.0.3
info:
  title: Hottub API
  description: |-
    API for hottub
  version: "1.0"
...
servers:
  - url: http://localhost/api
paths:
...
  /timers:
    get:
      tags:
        - timers
      summary: Get current timers of hottub
      description: Get current timers of hottub
      operationId: getTimers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimerEntity'
    post:
      tags:
        - timers
      summary: Add a new timer
      operationId: addTimer
      requestBody:
        description: Create a new timer
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Timer'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimerEntity'
  /timers/{timer-id}:
    put:
      tags:
        - timers
      summary: Update an existing timer
      description: Update an existing timer
      operationId: updateTimer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Timer'
        required: true
      parameters:
        - name: timer-id
          in: path
          description: ID of timer
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimerEntity'
    delete:
      tags:
        - timers
      summary: Delete an existing timer
      description: Delete an existing timer
      operationId: deleteTimer
      parameters:
        - name: timer-id
          in: path
          description: ID of timer
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successful operation
...
components:
  schemas:
...
    Settings:
      required:
        - temperatureLimit
        - temperatureDelta
        - heatingPanTemperatureLimit
        - circulationTimeCycle
        - lightsOn
        - debugMode
      type: object
      properties:
        temperatureLimit:
          type: integer
          minimum: 5
          maximum: 45
          description: Wanted temperature in hot tub
        temperatureDelta:
          type: integer
          description: Delta temperature difference between sensed temperature and actual temperature. Needed since temperature sensor is in the bottom, meaning temperatures will differ a few degrees from the sensor and the top of the hottub
        heatingPanTemperatureLimit:
          type: integer
          minimum: 15
          maximum: 60
          description: Maximum temperature allowed in heating pan until circulation pump should start operating
        circulationTimeCycle:
          type: integer
          description: Number of minutes between circulation pump operations, to make sure nothing freezes in winter and to stir around the water a bit
        lightsOn:
          type: boolean
          description: Whether the lights should be on or off
        debugMode:
          type: boolean
          description: Activate debug logs, which will log a lot more of operational information
    Timer:
      required:
        - settings
        - time
      type: object
      properties:
        settings:
          $ref: '#/components/schemas/Settings'
        time:
          type: string
          format: date-time
          description: Timestamp for when this settings should kick in
    TimerEntity:
      description: An existing timer entity
      allOf:     # Combines the BasicErrorModel and the inline model
        - $ref: '#/components/schemas/Timer'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              format: uuid
...